requirements.txt
asgiref==3.8.1
boto3==1.34.143
botocore==1.34.143
cloudinary==1.36.0
dj-database-url==2.2.0
dj3-cloudinary-storage==0.0.6
Django==5.0.6
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-storages==1.14.4
gunicorn==22.0.0
jmespath==1.0.1
pillow==10.4.0
psycopg==3.2.1
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.2
sqlparse==0.5.0
urllib3==1.26.19
whitenoise==6.7.0
